AFNetworking eliminated. All REST petitions made with OSWebRequest

+++ OSLibrary/OSCrypto.m
@@ -17,7 +17,7 @@
 + (NSString *)md5:(NSString *)str { 
   const char *cStr = [str UTF8String];
   unsigned char result[CC_MD5_DIGEST_LENGTH]; 
-  CC_MD5(cStr, strlen(cStr), result); 
+  CC_MD5(cStr, (unsigned int)strlen(cStr), result);
   return [NSString stringWithFormat:
       @"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
       result[0], result[1], result[2], result[3],
